+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [5.19.0] - 2026-08-03
+
+_Stable release based on [5.19.0-rc.1]._
+
+## [5.19.0-rc.1] - 2026-08-03
+
+### Changed
+
+- Change association history timestamps to "$date by $name" format.
+- Show the latest time of the association history time frame.
+- Sort association history entries within a time frame ascending.
+
 ## [5.18.14] - 2026-08-03
 
 ### Fixed
@@ -1706,6 +1718,8 @@ _Stable release based on [0.1.0-rc.1]._
 
 ## [0.0.0] - 2024-09-12
 
+[5.19.0]: https://https://github.com/internetguru/laravel-common/compare/v5.18.14...v5.19.0
+[5.19.0-rc.1]: https://github.com/internetguru/laravel-common/releases/tag/v5.18.14
 [5.18.14]: https://https://github.com/internetguru/laravel-common/compare/v5.18.13...v5.18.14
 [5.18.13]: https://https://github.com/internetguru/laravel-common/compare/v5.18.12...v5.18.13
 [5.18.12]: https://https://github.com/internetguru/laravel-common/compare/v5.18.11...v5.18.12